# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skill Eval Platform API",
    "description": "Backend service for the unified evaluation question library.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/scenes": {
      "get": {
        "tags": [
          "scenes"
        ],
        "summary": "List Scenes",
        "operationId": "list_scenes_api_scenes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneListResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          }
        }
      },
      "post": {
        "tags": [
          "scenes"
        ],
        "summary": "Create Scene",
        "operationId": "create_scene_api_scenes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneView"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "409": {
            "description": "同名场景已存在"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenes/{scene_id}": {
      "get": {
        "tags": [
          "scenes"
        ],
        "summary": "Get Scene Status",
        "operationId": "get_scene_status_api_scenes__scene_id__get",
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "场景不存在"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenes/{scene_id}/credentials": {
      "post": {
        "tags": [
          "scenes"
        ],
        "summary": "Issue Credential",
        "operationId": "issue_credential_api_scenes__scene_id__credentials_post",
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scene Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneCredentialIssueRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneCredentialIssuedView"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "场景不存在"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenes/{scene_id}/credentials/rotation": {
      "post": {
        "tags": [
          "scenes"
        ],
        "summary": "Rotate Credentials",
        "operationId": "rotate_credentials_api_scenes__scene_id__credentials_rotation_post",
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scene Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneCredentialIssueRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneCredentialIssuedView"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "场景不存在"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenes/{scene_id}/credentials/{credential_id}": {
      "delete": {
        "tags": [
          "scenes"
        ],
        "summary": "Revoke Credential",
        "operationId": "revoke_credential_api_scenes__scene_id__credentials__credential_id__delete",
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scene Id"
            }
          },
          {
            "name": "credential_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Credential Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneCredentialStatusView"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "场景凭证不存在"
          },
          "409": {
            "description": "凭证已被撤销"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SceneCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SceneCreateRequest"
      },
      "SceneCredentialIssueRequest": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SceneCredentialIssueRequest"
      },
      "SceneCredentialIssuedView": {
        "properties": {
          "credential_id": {
            "type": "string",
            "title": "Credential Id"
          },
          "scene_id": {
            "type": "string",
            "title": "Scene Id"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "credential_id",
          "scene_id",
          "token",
          "created_at"
        ],
        "title": "SceneCredentialIssuedView",
        "description": "Returned exactly once at issue/rotate time; plaintext is never stored."
      },
      "SceneCredentialStatusView": {
        "properties": {
          "credential_id": {
            "type": "string",
            "title": "Credential Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "revoked"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "revoked_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked Reason"
          }
        },
        "type": "object",
        "required": [
          "credential_id",
          "label",
          "status",
          "created_at",
          "last_used_at",
          "revoked_at",
          "revoked_reason"
        ],
        "title": "SceneCredentialStatusView"
      },
      "SceneListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SceneView"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SceneListResponse"
      },
      "SceneStatusResponse": {
        "properties": {
          "scene": {
            "$ref": "#/components/schemas/SceneView"
          },
          "credentials": {
            "items": {
              "$ref": "#/components/schemas/SceneCredentialStatusView"
            },
            "type": "array",
            "title": "Credentials"
          }
        },
        "type": "object",
        "required": [
          "scene",
          "credentials"
        ],
        "title": "SceneStatusResponse"
      },
      "SceneView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count"
          },
          "active_credential_count": {
            "type": "integer",
            "title": "Active Credential Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "question_count",
          "active_credential_count"
        ],
        "title": "SceneView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}